requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
python-dotenv==1.0.0
aiofiles==23.2.1
